Feat: 관리자 기능 구현

See merge request s12-webmobile2-sub1/S12P11A703!82

+++ backend/docshund/src/main/java/com/ssafy/docshund/domain/docs/repository/TranslatedDocumentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.ssafy.docshund.domain.docs.entity.Status;
 import com.ssafy.docshund.domain.docs.entity.TranslatedDocument;
 
-public interface TranslatedDocumentRepository extends JpaRepository<TranslatedDocument, Integer> {
+public interface TranslatedDocumentRepository extends JpaRepository<TranslatedDocument, Long> {
 
 	// 특정 원본 문서(docsId)에 속한 번역 문서 조회
 	List<TranslatedDocument> findByOriginDocument_Document_DocsIdAndStatus(Integer docsId, Status status);
